cast integers explicitly
Former-commit-id: e9e414d66714cd92b92c138801af358b086893fe

+++ Source/ThreadLevelScheduler.cpp
@@ -53,9 +53,9 @@ void ThreadLevelScheduler::createSchedule(const Task::List& tasks, const Edges& 
 			if (mSortTaskTypes)
 				sortTasksByType(levels[level].begin(), levels[level].end());
 			// Distribute tasks of one level evenly between threads
-			for (int thread = 0; thread < mNumThreads; thread++) {
-				int start = levels[level].size() * thread / mNumThreads;
-				int end = levels[level].size() * (thread + 1) / mNumThreads;
+			for (Int thread = 0; thread < mNumThreads; thread++) {
+				Int start = static_cast<Int>(levels[level].size()) * thread / mNumThreads;
+				Int end = static_cast<Int>(levels[level].size()) * (thread + 1) / mNumThreads;
 				for (int idx = start; idx != end; idx++)
 					scheduleTask(thread, levels[level][idx]);
 			}
